factored out srunit_servers;stateful_compute_test passes
